+++ dependencies/reactphysics3d/0.7.1/src/collision/TriangleMesh.h
@@ -92,7 +92,7 @@ inline TriangleVertexArray* TriangleMesh::getSubpart(uint indexSubpart) const {
  * @return The number of sub-parts of the mesh
  */
 inline uint TriangleMesh::getNbSubparts() const {
-    return mTriangleArrays.size();
+    return static_cast<uint>(mTriangleArrays.size());
 }
 
 }
